+++ src/main/java/seedu/address/logic/parser/DeleteOrderCommandParser.java
@@ -4,7 +4,10 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.DeleteOrderCommand;
+import static seedu.address.logic.parser.CliSyntax.DELETE_PREFIX_ORDER;
 import seedu.address.logic.parser.exceptions.ParseException;
+
+import java.util.stream.Stream;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
@@ -18,7 +21,17 @@ public class DeleteOrderCommandParser implements Parser<DeleteOrderCommand> {
      */
     public DeleteOrderCommand parse(String args) throws ParseException {
         try {
-            Index index = ParserUtil.parseIndex(args);
+            ArgumentMultimap argMultimap =
+                    ArgumentTokenizer.tokenize(args, DELETE_PREFIX_ORDER);
+
+            if (!arePrefixesPresent(argMultimap,
+                    DELETE_PREFIX_ORDER)
+                    || !argMultimap.getPreamble().isEmpty()) {
+                throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteOrderCommand.MESSAGE_USAGE));
+            }
+
+            Index index = ParserUtil.parseOrderIndex(argMultimap.getValue(DELETE_PREFIX_ORDER).get());
+
             return new DeleteOrderCommand(index);
         } catch (ParseException pe) {
             throw new ParseException(
@@ -26,5 +39,12 @@ public class DeleteOrderCommandParser implements Parser<DeleteOrderCommand> {
         }
     }
 
+    /**
+     * Returns true if none of the prefixes contains empty {@code Optional} values in the given
+     * {@code ArgumentMultimap}.
+     */
+    private static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix... prefixes) {
+        return Stream.of(prefixes).allMatch(prefix -> argumentMultimap.getValue(prefix).isPresent());
+    }
 }
